+++ macros/CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   field.
 - Add derive macros for `KeyTag`, `PrimaryKeyTag`, `SerializeKey` and `DeserializeKey`. They can
   only be used with newtypes.
+- `Introspectable` now also supports newtypes.
 
 ### Changed